check for MAX_LINE_LEN in main.c:chomp to not overflow the buffer
git-svn-id: http://dzen.googlecode.com/svn/trunk@90 f2baff5b-bf2c-0410-a398-912abdc3d8b2

+++ dzen.h
@@ -96,7 +96,7 @@ struct DZEN {
 	Bool ispersistent;
 	Bool tsupdate;
 	unsigned long timeout;
-	int cur_line;
+	long cur_line;
 	int ret_val;
 
 	/* should always be 0 if DZEN_XINERAMA not defined */
